requirements.txt
anyio==3.3.4
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
neo4j==4.3.7
pydantic==1.8.2
python-dotenv==0.19.1
pytz==2021.3
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
